Avoid race-condition in processRecord (fixes #1).
This race condition had the effect that the bo record’s value sometimes
was not reset to zero when the execution of the command had finished.

This happened because the check in the processRecord method for whether
execution of the command had finished would sometimes fail even though
execution had actually finished. This happened when the scheduled
callback was executed before the thread scheduling the callback had
finished, which due to scheduling could happen on any system, but it
was particularly likely to happen on fast multi-core systems.

This would not lead to an undefined state, but the record’s value would
only be reset to zero the next time the record was processed explicitly,
which could be inconvenient when the record’s value was used to check
whether the command was still running.