Avoid accumulating zombie processes (fixes #3).
When using the no-wait flag for a command, the IOC would never wait for
the child process created for executing this command to terminate. This
meant that when this process terminated, it would stay as a zombie
process until the whole IOC was terminated. Thus, many Zombie processes
could accumulate over time, consuming system resources.

This is now fixed by ensuring that waitpid() is always called for the
child processes that are created by the device support, even if we are
not interested in the child’s exit status. This call to waitpid()
happens in a background thread in order to not change semantics and
return immediately when the no-wait flag is set.