Bugfix: Do not block in record processing (fixes #5).
With #4, a bug was introduced that causes processing of the bo record
triggering the execution of the command to block until the execution of
the command complete, when the no-wait flag is set.

The cause was that Command::run created a future using std::async, but
this future was destroyed when the function returned. However, the
destructor of std::future may block when the future has been created by
std::async and the shared state is not available yet.

This is fixed now by keeping the future and only destroying it on the
next call to Command::run, when the shared state is available.