Avoid use of std::async (closes #9).
It seems like std::async uses a limited number of threads on some
platforms (which is perfectly legal according to the C++ standard). This
means that a long-running command might cause other commands from being
executed when the threads for that long-running command have executed
the thread pool used by std::async.

By using our own thread pool, that grows as needed, we can avoid waiting
for a thread to become available.