Close unused FDs in child process (closes #9).
Before executing a new process after forking, we have to close all file
descriptors that are not explicitly needed in the child process.
Otherwise, due to the IOC being multi-threaded, the file descriptors
associated with the pipes for other child processes might leak into the
child process, causing these pipes to not be closed when the child
process for which they were intended quits.

This also helps ensure that the executed program does not get access to
resources to which it should not have access (e.g. sockets, etc.).

In addition to this change, we ensure that the file descriptors for
stdin, stdout, and stderr are always available in the launched program,
binding them to /dev/null if we do not have a pipe for them. If we did
not do this and the executed program opened a file, it might get one of
these file descriptors assigned, which could lead to this file being
used for input or output where it should not (e.g. when the program
writes diagnostic information to stderr, assuming that it is either an
invalid (closed) file descriptor or that it actually refers to the right
target for diagnostic messages.